+++ crates/stryx_rules/src/flows/xss_via_dangerously_set_inner_html.rs
@@ -22,9 +22,9 @@ use std::collections::HashMap;
 use stryx_ast::{
     Visit,
     ast::{
-        Argument, ArrowFunctionExpression, BindingPattern, CallExpression, ChainElement,
-        Expression, Function, JSXAttribute, JSXAttributeValue, JSXExpression, ObjectExpression,
-        ObjectPropertyKind, PropertyKey, VariableDeclarator,
+        Argument, ArrowFunctionExpression, AssignmentExpression, AssignmentTarget, BindingPattern,
+        CallExpression, ChainElement, Expression, Function, JSXAttribute, JSXAttributeValue,
+        JSXExpression, ObjectExpression, ObjectPropertyKind, PropertyKey, VariableDeclarator,
     },
     to_span,
 };
@@ -177,6 +177,7 @@ impl XssVisitor {
                 ChainElement::PrivateFieldExpression(m) => self.expr_taint(&m.object),
                 ChainElement::TSNonNullExpression(t) => self.expr_taint(&t.expression),
             },
+            Expression::AssignmentExpression(a) => self.expr_taint(&a.right),
             _ => false,
         }
     }
@@ -278,6 +279,18 @@ impl<'a> Visit<'a> for XssVisitor {
     fn visit_jsx_attribute(&mut self, attr: &JSXAttribute<'a>) {
         self.check_jsx_attribute(attr);
         stryx_ast::walk::walk_jsx_attribute(self, attr);
+    }
+
+    fn visit_assignment_expression(&mut self, a: &AssignmentExpression<'a>) {
+        let rhs_tainted = self.expr_taint(&a.right);
+        if let AssignmentTarget::AssignmentTargetIdentifier(id) = &a.left {
+            if rhs_tainted {
+                self.taint(id.name.to_string());
+            } else if let Some(scope) = self.scopes.last_mut() {
+                scope.remove(id.name.as_str());
+            }
+        }
+        self.visit_expression(&a.right);
     }
 }